pyproject.toml
[tool.poetry]
name = "ai-gemini-chatbot"
version = "0.1.0"
description = ""
authors = ["danielAdama <adamadaniel321@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-generativeai = "^0.8.2"
jupyterlab = "^4.2.5"
python-dotenv = "^1.0.1"
redis = "^5.1.0"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
